Presenter refactor
