+++ app/src/main/java/com/starwarscharacters/ui/about/AboutPresenter.kt
@@ -1,6 +1,6 @@
 package com.starwarscharacters.ui.about
 
-import com.starwarscharacters.Presenter
+import com.starwarscharacters.ui.Presenter
 
 object AboutPresenter: Presenter<AboutScreen?>() {
 }
